Created customizable ComputingNodesGenerator + small fixes
