+++ PureEdgeSim/com/mechalikh/pureedgesim/datacentersmanager/DataCentersManager.java
@@ -20,6 +20,7 @@
  **/
 package com.mechalikh.pureedgesim.datacentersmanager;
 
+import java.lang.Class;
 import java.lang.reflect.Constructor;
 
 import com.mechalikh.pureedgesim.locationmanager.MobilityModel;
@@ -48,7 +49,7 @@ public class DataCentersManager {
 	 * files.
 	 * 
 	 * @see #generateComputingNodes()
-	 * @see com.mechalikh.pureedgesim.datacentersmanager.ComputingNodesGenerator
+	 * @see com.mechalikh.pureedgesim.datacentersmanager.DefaultComputingNodesGenerator
 	 */
 	protected ComputingNodesGenerator computingNodesGenerator;
 
@@ -62,21 +63,28 @@ public class DataCentersManager {
 	/**
 	 * Initializes the DataCentersManager
 	 *
-	 * @param simulationManager  The simulation Manager
-	 * @param mobilityModelClass The mobility model that will be used in the
-	 *                           simulation
-	 * @param computingNodeClass The computing node class that will be used to
-	 *                           generate computing resources
-	 * @param topologyCreator
+	 * @param simulationManager  			The simulation manager
+	 * @param mobilityModelClass 			The mobility model class that will be used in the
+	 *                           			simulation
+	 * @param computingNodeClass 			The computing node class that will be used to
+	 *                           			generate computing resources
+	 * @param computingNodesGeneratorClass	The computing node generator class that will be
+	 * 										used to generate all resources from the XML
+	 * 										files
+	 * @param topologyCreatorClass			The topology creator class that will be used to
+	 * 										generate the network topology
 	 */
-	public DataCentersManager(SimulationManager simulationManager, Class<? extends MobilityModel> mobilityModelClass,
-			Class<? extends ComputingNode> computingNodeClass, Class<? extends TopologyCreator> topologyCreatorClass) {
+	public DataCentersManager(SimulationManager simulationManager, 
+			Class<? extends MobilityModel> mobilityModelClass,
+			Class<? extends ComputingNode> computingNodeClass, 
+			Class<? extends ComputingNodesGenerator> computingNodesGeneratorClass,
+			Class<? extends TopologyCreator> topologyCreatorClass) {
 		this.simulationManager = simulationManager;
 		// Add this to the simulation manager and submit computing nodes to broker
 		simulationManager.setDataCentersManager(this);
 
-		// Generate all data centers, servers, an devices
-		generateComputingNodes(mobilityModelClass, computingNodeClass);
+		// Generate all data centers, servers, and devices
+		generateComputingNodes(mobilityModelClass, computingNodeClass, computingNodesGeneratorClass);
 
 		// Generate topology
 		createTopology(topologyCreatorClass);
@@ -85,14 +93,24 @@ public class DataCentersManager {
 	/**
 	 * Generates all computing nodes.
 	 * 
-	 * @param computingNodeClass
 	 * @param mobilityModelClass
+	 * @param computingNodeClass
+	 * @param computingNodesGeneratorClass
 	 */
 	protected void generateComputingNodes(Class<? extends MobilityModel> mobilityModelClass,
-			Class<? extends ComputingNode> computingNodeClass) {
+			Class<? extends ComputingNode> computingNodeClass, 
+			Class<? extends ComputingNodesGenerator> computingNodesGeneratorClass) {
 		SimLog.println("%s - Generating computing nodes...",this.getClass().getSimpleName());
-		computingNodesGenerator = new ComputingNodesGenerator(simulationManager, mobilityModelClass,
-				computingNodeClass);
+		Constructor<?> computingNodesGeneratorConstructor;
+		try {
+			computingNodesGeneratorConstructor = computingNodesGeneratorClass.getConstructor(SimulationManager.class,
+					Class.class, Class.class);
+
+			computingNodesGenerator = (ComputingNodesGenerator) computingNodesGeneratorConstructor.newInstance(simulationManager,
+					mobilityModelClass, computingNodeClass);
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
 		computingNodesGenerator.generateDatacentersAndDevices();
 	}
