+++ PureEdgeSim/com/mechalikh/pureedgesim/simulationmanager/SimulationThread.java
@@ -198,7 +198,7 @@ public class SimulationThread {
 		// Generate all data centers, servers, an devices
 		SimLog.println(this.getClass().getSimpleName() + " - Initializing the Datacenters Manager Module...");
 		new DataCentersManager(simulationManager, simulation.mobilityModel, simulation.computingNode,
-				simulation.topologyCreator);
+				simulation.computingNodesGenerator, simulation.topologyCreator);
 
 		// Generate tasks list
 		SimLog.println(this.getClass().getSimpleName() + " - Initializing the Task Generator...");
